fix: put dashboard routes inside dashboard components fix #8913
